requirements.txt
python==3.10.8
ipykernel 
keras==2.11.0
matplotlib==3.6.2
numpy==1.23.5
opencv-python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
sklearn==0.0.post1
streamlit==1.15.2
streamlit-drawable-canvas==0.9.2
tensorflow==2.11.0
